+++ CHANGELOG.md
@@ -8,6 +8,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Added better previews for urls shared in notes.
+
 ## [0.1 (57)] - 2023-07-17Z
 
 - Fixed an issue where Nos couldn't find the user's key on startup.